Move action table to components folder and add context text to new view
