+++ frontend/src/components/AnnualActionableCanteensTable.vue
